Split schema_repository.go back out from schema.go
Keep the repository interface in its own file to match the existing
convention used by RelationsRepository and ResourceRepository.
